pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "kronoterm-cloud-relay"
version = "0.0.22"
readme = "README.md"
description = "Relay between kronoterm cloud API and local network API"
authors = [{name = "Matevž", email = "matevzlesko@gmail.com"},]


license = "GPL-3.0-or-later"

requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.115.12",
    "python-dotenv>=1.0.1",
    "kronoterm-cloud-api>=0.1.16"
]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/LeskoIam/kronoterm_cloud_relay"
Issues = "https://github.com/LeskoIam/kronoterm_cloud_relay/issues"


[tool.bumpver]
current_version = "0.0.22"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "Bump version {old_version} -> {new_version}."
tag_message = "v{new_version}"
tag_scope = "default"
pre_commit_hook = ""
post_commit_hook = ""
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]
"src/kronoterm_cloud_relay.py" = [
    '__version__ = "{version}"',
]
"docker_release.bat" = [
    'docker build -t leskoiam/kronoterm_cloud_relay:{version} .',
    'docker tag leskoiam/kronoterm_cloud_relay:{version} leskoiam/kronoterm_cloud_relay:latest',
    'docker push leskoiam/kronoterm_cloud_relay:{version}'
]


[tool.ruff]
target-version = "py312"
line-length = 121
indent-width = 4

# Exclude files
extend-exclude = ["setup.py"]

[tool.ruff.lint]
# select = ["ALL"]
select = [
    # pycodestyle errors
    "E",
    # pycodestyle warnings
    "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # pep8-naming
    "N",
    # mccabe
    "C901",
    # pydocstyle
    "D",
    # flake8-pytest-style
    "PT"
]

ignore = [
    # Missing docstring in public module
    "D100",
    # Missing docstring in public class
    "D101",
    # Missing docstring in public package
    "D104",
    # Missing docstring in magic method
    "D105",
    # No blank lines allowed after function docstring
    "D202",
    # 1 blank line required before class docstring
    "D203",
    # 1 blank line required between summary line and description
    "D205",
    # Multi-line docstring summary should start at the second line
    "D213",
    # First line should end with a period
    "D400",
    # First line of docstring should be in imperative mood
    "D401",
    #  Multi-line docstring summary should start at the first line
    "D212",
    # First line should end with a period, question mark, or exclamation point
    "D415",
    # Missing blank line after last section
    "D413",
    # Missing dashed underline after section
    "D407",
    # Section name should end with a newline
    "D406",
    # Unnecessary open mode parameters (explicit is better than implicit)
    "UP015",
    # Local variable is assigned to but never used
    "F841",
    # Use context handler for opening files
    "SIM115",
    # `scope='function'` is implied in `@pytest.fixture()`
    "PT003",
    # Exception should be named with an Error suffix
    "N818",
    # Use ternary operator instead of `if`-`else`-block
    "SIM108"

]

# Avoid trying to fix flake8-bugbear (`B`) violations
unfixable = ["B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"


[tool.pytest.ini_options]
pythonpath = [
    "src",
]